+++ CHANGES.md
@@ -67,6 +67,7 @@ Changes in 2.9
 - Declare $error, $note, $warning globals in ErrorMessage.fnc.php & program wide
 - Add Student Assignments program + related SQL + assets/AssignmentsFiles/ folder
 - Use ISO date format (YYYY-MM-DD), always in Date.php & program wide
+- Add ProperDateTime() function in Date.php
 
 Changes in 2.8.11
 -----------------